HaskellBook.Examples

Documentation

module HaskellBook.Examples.UserInfo

data Name

Constructors

Name String

data Salutation

Constructors

Salutation String

data GreetingMessage

A GreetingMessage contains all of the information needed to generate a greeting using formatMessage. You can get a default greeting without attribution from defaultMessage. This makes it convenient to use record update syntax to construct a new greeting:

>>> formatMessage defaultMessage { greetingFrom = [ Name "A Haskeller"] }
"Hello Friend, from: A Haskeller"

Constructors

GreetingMessage

Fields

greetingSalutation :: Salutation

A Salutation, like "Hello"

greetingTo :: Name

Name of the person that should be greeted

greetingFrom :: [Name]

Names of the people who are sending the greeting

defaultMessage :: GreetingMessage

A default greeting message that isn't attributed to anyone:

GreetingMessage
  { greetingSalutation = Salutation "Hello"
  , greetingTo = Name "Friend"
  , greetingFrom = []
  }

formatMessage :: GreetingMessage -> String

data SortedList

makeSortedList :: [Int] -> Maybe SortedList

minimum :: SortedList -> Int